Import position changes from DOLSharp (https://github.com/Dawn-of-Light/DOLSharp/pull/472)
Squashed commit of the following:

commit f4d1210d8f66565a6423c646f3e2fff0d26aa7bb
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 20:28:26 2024 -0400

    improv: remove old Geometry types

commit 0dd00552b0c9b1a0dbbbf592fb299315ceb94c84
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 20:17:55 2024 -0400

    fix: fix OnPlayerMove not being called

commit 5eaf744f058e1703cf8c9a1ec7449722aaed707e
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 20:03:24 2024 -0400

    fix: tentatively fix failure to interrupt threads on exit

commit 4a81eaff7adcd6039db9f08983481a1be6b0d619
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 18:43:42 2024 -0400

    fix: fix exception in format in GameTimer.cs

commit a7b55b7ac39dbf467d836b15fe0084bb24a25044
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 18:43:18 2024 -0400

    improv: cache Region in Position

commit fd55949a3b980fb7f49fa40c53d5736406c4eec9
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 18:27:14 2024 -0400

    Fix Areas, I think

commit 507a7288d3d8b5a56326b34f70695a5d4d20b2cd
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 18:05:23 2024 -0400

    Fix overflow when calculating very long distances

commit aa3e1e5754ed30fa06eaf6efd3cc47edba339a81
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 14 17:11:03 2024 -0400

    improv: improve "Zone does not exist" error log in AddToWorld

commit dc292cd6e7cc64d9440b3a27b3292ad74cd669f0
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Sun Jun 9 18:01:06 2024 -0400

    Fix some issues

commit f59ed078a737b4697408986b4260fe5d88b9cef7
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Sun Jun 9 15:59:17 2024 -0400

    Apply position changes to Gondwana code

commit b0d12d2395c5909a31f6dacd2e86e021c280afd5
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Sat Jun 8 16:14:51 2024 -0400

    Last batch of importing the DOLSharp positions - still need to change Gondwana-specific code

commit 966394b2945c1d16934ae9ab5f8381822f3663dd
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Fri Jun 7 19:23:28 2024 -0400

    More of the position changes, still not finished

commit a873193a91bec1b284e9e03ac7e44c79e033fea7
Author: Amber Ehrlich <miuna.oshino@gmail.com>
Date:   Sun Jun 2 19:58:37 2024 -0400

    positions, part 1

+++ GameServer/commands/gmcommands/area.cs
@@ -77,9 +77,9 @@ namespace DOL.GS.Commands
                         }
                         area.Sound = byte.Parse(args[6]);
                         area.Region = client.Player.CurrentRegionID;
-                        area.X = (int)client.Player.Position.X;
-                        area.Y = (int)client.Player.Position.Y;
-                        area.Z = (int)client.Player.Position.Z;
+                        area.X = client.Player.Position.X;
+                        area.Y = client.Player.Position.Y;
+                        area.Z = client.Player.Position.Z;
                         area.ObjectId = area.Description;
 
                         if (args.Length == 8 && bool.TryParse(args[7], out bool canVol))
